0-linear.c

+++ 0x1E-search_algorithms/0-linear.c
@@ -12,16 +12,16 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	size_t x;
+	int x;
 
 	if (array == NULL)
 	{
 		return (-1);
 	}
 
-	for (x = 0; x < size; x++)
+	for (x = 0; x < (int)size; x++)
 	{
-		printf("Value checked array[%li] = [%i]\n", x, array[x]);
+		printf("Value checked array[%u] = [%d]\n", x, array[x]);
 		if (array[x] == value)
 		{
 			return (x);
